+++ src/graph.h
@@ -9,10 +9,10 @@ using namespace std;
 class Graph {
     public:
     Graph(set<int> v);
-    vector<vector<int>> adj;
+    vector<vector<long long>> adj;
     vector<vector<int>> edges;
     set<int> V;
-    void add_edge(int from, int to, int weight);
+    void add_edge(int from, int to, long long weight);
 };
 
 struct SSSP_Result {
